move the custom parser from flockmtl/core/parser into flockmtl/custom_parser
