+++ src/include/flockmtl/custom_parser/query/prompt_parser.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "flockmtl/common.hpp"
-#include "flockmtl/core/parser/query_statements.hpp"
-#include "flockmtl/core/parser/tokenizer.hpp"
+#include "flockmtl/custom_parser/query_statements.hpp"
+#include "flockmtl/custom_parser/tokenizer.hpp"
 
 #include <memory>
 #include <string>
@@ -10,47 +10,36 @@
 
 namespace flockmtl {
 
-namespace core {
 // Specific prompt statements
 class CreatePromptStatement : public QueryStatement {
 public:
-    CreatePromptStatement() {
-        type = StatementType::CREATE_PROMPT;
-    }
+    CreatePromptStatement() { type = StatementType::CREATE_PROMPT; }
     std::string prompt_name;
     std::string prompt;
 };
 
 class DeletePromptStatement : public QueryStatement {
 public:
-    DeletePromptStatement() {
-        type = StatementType::DELETE_PROMPT;
-    }
+    DeletePromptStatement() { type = StatementType::DELETE_PROMPT; }
     std::string prompt_name;
 };
 
 class UpdatePromptStatement : public QueryStatement {
 public:
-    UpdatePromptStatement() {
-        type = StatementType::UPDATE_PROMPT;
-    }
+    UpdatePromptStatement() { type = StatementType::UPDATE_PROMPT; }
     std::string prompt_name;
     std::string new_prompt;
 };
 
 class GetPromptStatement : public QueryStatement {
 public:
-    GetPromptStatement() {
-        type = StatementType::GET_PROMPT;
-    }
+    GetPromptStatement() { type = StatementType::GET_PROMPT; }
     std::string prompt_name;
 };
 
 class GetAllPromptStatement : public QueryStatement {
 public:
-    GetAllPromptStatement() {
-        type = StatementType::GET_ALL_PROMPT;
-    }
+    GetAllPromptStatement() { type = StatementType::GET_ALL_PROMPT; }
 };
 
 class PromptParser {
@@ -64,7 +53,5 @@ private:
     void ParseUpdatePrompt(Tokenizer &tokenizer, std::unique_ptr<QueryStatement> &statement);
     void ParseGetPrompt(Tokenizer &tokenizer, std::unique_ptr<QueryStatement> &statement);
 };
-
-} // namespace core
 
 } // namespace flockmtl
