+++ src/include/flockmtl/custom_parser/query/secret_parser.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "flockmtl/common.hpp"
-#include "flockmtl/core/parser/query_statements.hpp"
-#include "flockmtl/core/parser/tokenizer.hpp"
+#include "flockmtl/custom_parser/query_statements.hpp"
+#include "flockmtl/custom_parser/tokenizer.hpp"
 
 #include <memory>
 #include <string>
@@ -10,48 +10,36 @@
 
 namespace flockmtl {
 
-namespace core {
-
 // Specific SECRET statements
 class CreateSecretStatement : public QueryStatement {
 public:
-    CreateSecretStatement() {
-        type = StatementType::CREATE_SECRET;
-    }
+    CreateSecretStatement() { type = StatementType::CREATE_SECRET; }
     std::string provider;
     std::string secret;
 };
 
 class DeleteSecretStatement : public QueryStatement {
 public:
-    DeleteSecretStatement() {
-        type = StatementType::DELETE_SECRET;
-    }
+    DeleteSecretStatement() { type = StatementType::DELETE_SECRET; }
     std::string provider;
 };
 
 class UpdateSecretStatement : public QueryStatement {
 public:
-    UpdateSecretStatement() {
-        type = StatementType::UPDATE_SECRET;
-    }
+    UpdateSecretStatement() { type = StatementType::UPDATE_SECRET; }
     std::string provider;
     std::string secret;
 };
 
 class GetSecretStatement : public QueryStatement {
 public:
-    GetSecretStatement() {
-        type = StatementType::GET_SECRET;
-    }
+    GetSecretStatement() { type = StatementType::GET_SECRET; }
     std::string provider;
 };
 
 class GetAllSecretStatement : public QueryStatement {
 public:
-    GetAllSecretStatement() {
-        type = StatementType::GET_ALL_SECRET;
-    }
+    GetAllSecretStatement() { type = StatementType::GET_ALL_SECRET; }
     std::string provider;
 };
 
@@ -66,7 +54,5 @@ private:
     void ParseUpdateSecret(Tokenizer &tokenizer, std::unique_ptr<QueryStatement> &statement);
     void ParseGetSecret(Tokenizer &tokenizer, std::unique_ptr<QueryStatement> &statement);
 };
-
-} // namespace core
 
 } // namespace flockmtl
